fix: successor is updated in time after failure
Kill channel of node is properly used now (it's closed in order to notify all goroutines). The same is done with RpcRing quit channel.

A custom ID can be provided to create a node. If so, such ID isn't hashed.

Fixed bug in utils.RepeatAction(): it never stopped when kill channel was closed.

Tests for joining to a ring of 2 and removing from a ring of 4.

Logs to file.

Refs: #5